+++ config/config_test.go
@@ -96,7 +96,7 @@ func (suite *ConfigTestSuite) TestValidateConfDefault() {
 	assert.Equal(suite.T(), "memory", suite.ConfGorushDefault.Stat.Engine)
 	assert.Equal(suite.T(), "localhost:6379", suite.ConfGorushDefault.Stat.Redis.Addr)
 	assert.Equal(suite.T(), "", suite.ConfGorushDefault.Stat.Redis.Password)
-	assert.Equal(suite.T(), int64(0), suite.ConfGorushDefault.Stat.Redis.DB)
+	assert.Equal(suite.T(), 0, suite.ConfGorushDefault.Stat.Redis.DB)
 
 	assert.Equal(suite.T(), "gorush.db", suite.ConfGorushDefault.Stat.BoltDB.Path)
 	assert.Equal(suite.T(), "gorush", suite.ConfGorushDefault.Stat.BoltDB.Bucket)
@@ -149,7 +149,7 @@ func (suite *ConfigTestSuite) TestValidateConf() {
 	assert.Equal(suite.T(), "memory", suite.ConfGorush.Stat.Engine)
 	assert.Equal(suite.T(), "localhost:6379", suite.ConfGorush.Stat.Redis.Addr)
 	assert.Equal(suite.T(), "", suite.ConfGorush.Stat.Redis.Password)
-	assert.Equal(suite.T(), int64(0), suite.ConfGorush.Stat.Redis.DB)
+	assert.Equal(suite.T(), 0, suite.ConfGorush.Stat.Redis.DB)
 
 	assert.Equal(suite.T(), "gorush.db", suite.ConfGorush.Stat.BoltDB.Path)
 	assert.Equal(suite.T(), "gorush", suite.ConfGorush.Stat.BoltDB.Bucket)
